결제 준비 단계
주문번호 생성 로직 중
시퀀스에서 주문번호 받을 때 받는 타입을
int -> long으로 수정

+++ honeyComboFactory/src/main/java/zic/honeyComboFactory/biz/purchase/impl/OraclePurchaseDAOMybatis.java
@@ -77,7 +77,7 @@ public class OraclePurchaseDAOMybatis { // 주문 기능 - Oracle DB
 	}
 
 	// 시퀀스 값 조회(주문번호 생성용)
-	public int getNextPurchaseSequence() {
+	public long getNextPurchaseSequence() {
 		return mybatis.selectOne("PurchaseDAO.GETNEXTPURCHASENUMBER");
 	}
 }
